requirements.txt
Flask
peewee
Flask-Login
weasyprint
Werkzeug
python-dotenv